# portal/public/openapi/lumbre.openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Lumbre API",
    "version": "1.0.0",
    "description": "Contract for the Lumbre system under test. Carts are scoped to an anonymous session or authenticated account; authentication and protected business mutations remain limited to development and test until production email delivery is configured."
  },
  "servers": [{ "url": "/", "description": "Current Lumbre environment" }],
...
  "paths": {
...
    "/api/account": {
      "get": {
        "operationId": "getAccount",
        "tags": ["Identity"],
        "security": [{}, { "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Current account, or null for an anonymous request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AccountResponse" } }
            }
          }
        }
      }
    },
    "/api/account/preferences": {
      "get": {
        "operationId": "getMembershipPreferences",
        "tags": ["Identity"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/MembershipPreferencesResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" }
        }
      },
      "put": {
        "operationId": "updateMembershipPreferences",
        "tags": ["Identity"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPreferenceUpdateRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/MembershipPreferenceMutationResponse" },
          "201": { "$ref": "#/components/responses/MembershipPreferenceMutationResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/account/magic-link": {
      "post": {
        "operationId": "requestAccountMagicLink",
        "tags": ["Identity"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/MagicLinkRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "A passwordless verification link was accepted for delivery",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MagicLinkResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "503": { "$ref": "#/components/responses/AuthenticationUnavailableResponse" }
        }
      }
    },
    "/api/account/logout": {
      "post": {
        "operationId": "logoutAccount",
        "tags": ["Identity"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "The authenticated session was invalidated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/LogoutResponse" } }
            }
          },
          "503": { "$ref": "#/components/responses/AuthenticationUnavailableResponse" }
        }
      }
    },
    "/api/admin/accounts": {
      "get": {
        "operationId": "listAccounts",
        "tags": ["Identity"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Account summaries visible to administrators",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AccountCollection" } }
            }
          },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" },
          "503": { "$ref": "#/components/responses/AuthenticationUnavailableResponse" }
        }
      }
    },
    "/api/admin/products": {
      "get": {
        "operationId": "listAdministrativeProducts",
        "tags": ["Catalog"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/AdministrativeProductCollectionResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" }
        }
      },
      "post": {
        "operationId": "createAdministrativeProduct",
        "tags": ["Catalog"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AdministrativeProductCreateRequest" } }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AdministrativeProductMutationResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/admin/products/{id}": {
      "patch": {
        "operationId": "updateAdministrativeProduct",
        "tags": ["Catalog"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NumericResourceId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AdministrativeProductUpdateRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AdministrativeProductUpdateResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "409": { "$ref": "#/components/responses/ConflictResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/admin/events": {
      "get": {
        "operationId": "listAdministrativeEvents",
        "tags": ["Catalog"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/AdministrativeEventCollectionResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" }
        }
      },
      "post": {
        "operationId": "createAdministrativeEvent",
        "tags": ["Catalog"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AdministrativeEventCreateRequest" } }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/AdministrativeEventMutationResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/admin/events/{id}": {
      "patch": {
        "operationId": "updateAdministrativeEvent",
        "tags": ["Catalog"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/NumericResourceId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AdministrativeEventUpdateRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AdministrativeEventUpdateResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "409": { "$ref": "#/components/responses/ConflictResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/admin/audit-events": {
      "get": {
        "operationId": "listAdministrativeAuditEvents",
        "tags": ["Catalog"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/AdministrativeAuditCollectionResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "403": { "$ref": "#/components/responses/AuthorizationRequiredResponse" }
        }
      }
    },
    "/api/recipes": {
      "get": {
        "operationId": "listRecipes",
        "tags": ["Catalog"],
        "parameters": [
          { "$ref": "#/components/parameters/RecipeCategory" },
          { "$ref": "#/components/parameters/SearchQuery" }
        ],
        "responses": {
          "200": {
            "description": "Filtered recipe collection",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RecipeCollection" } }
            }
          }
        }
      }
    },
...
    "/api/cart/items": {
      "post": {
        "operationId": "addAnonymousCartItem",
        "tags": ["Commerce"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CartItemAddRequest" } }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/CartResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/cart/items/{productId}": {
      "patch": {
        "operationId": "updateAnonymousCartItem",
        "tags": ["Commerce"],
        "parameters": [{ "$ref": "#/components/parameters/ProductId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CartItemUpdateRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/CartResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      },
      "delete": {
        "operationId": "removeAnonymousCartItem",
        "tags": ["Commerce"],
        "parameters": [{ "$ref": "#/components/parameters/ProductId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/CartResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/orders": {
      "get": {
        "operationId": "listOrders",
        "tags": ["Commerce"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/OrderCollectionResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" }
        }
      },
      "post": {
        "operationId": "createOrder",
        "tags": ["Commerce"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/OrderCreateRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/OrderMutationResponse" },
          "201": { "$ref": "#/components/responses/OrderMutationResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "409": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "getOrder",
        "tags": ["Commerce"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/OrderId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/OrderResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" }
        }
      }
    },
    "/api/orders/{id}/payment": {
      "post": {
        "operationId": "payOrder",
        "tags": ["Commerce"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PaymentRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PaymentResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "402": { "$ref": "#/components/responses/PaymentResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "409": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/orders/{id}/checkout-session": {
      "post": {
        "operationId": "createHostedCheckoutSession",
        "tags": ["Commerce"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/HostedCheckoutResponse" },
          "201": { "$ref": "#/components/responses/HostedCheckoutResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "409": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "503": { "$ref": "#/components/responses/ProviderUnavailableResponse" }
        }
      }
    },
    "/api/payments/stripe/webhook": {
      "post": {
        "operationId": "receiveStripeWebhook",
        "tags": ["Commerce"],
        "parameters": [{ "$ref": "#/components/parameters/StripeSignature" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/StripeCheckoutEvent" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/ProviderEventResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "503": { "$ref": "#/components/responses/ProviderUnavailableResponse" }
        }
      }
    },
...
    "/api/events/{id}/reservations": {
      "post": {
        "operationId": "createEventReservation",
        "tags": ["Events"],
        "parameters": [{ "$ref": "#/components/parameters/EventId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ReservationCreateRequest" } }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/ReservationMutationResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "409": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/reservations": {
      "get": {
        "operationId": "listAccountReservations",
        "tags": ["Events"],
        "responses": {
          "200": { "$ref": "#/components/responses/ReservationCollectionResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" }
        }
      }
    },
    "/api/fire-presets": {
      "get": {
        "operationId": "listFirePlannerPresets",
        "tags": ["Planning"],
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/FirePresetCollectionResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" }
        }
      },
      "post": {
        "operationId": "saveFirePlannerPreset",
        "tags": ["Planning"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/FirePresetSaveRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/FirePresetMutationResponse" },
          "201": { "$ref": "#/components/responses/FirePresetMutationResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "409": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/fire-presets/sync": {
      "post": {
        "operationId": "syncLocalFirePlannerPresets",
        "tags": ["Planning"],
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/FirePresetSyncRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/FirePresetSyncResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" }
        }
      }
    },
    "/api/fire-presets/{id}": {
      "delete": {
        "operationId": "deleteFirePlannerPreset",
        "tags": ["Planning"],
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/FirePresetId" }],
        "responses": {
          "200": {
            "description": "The account-owned preset was deleted",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FirePresetDeleteResult" } }
            }
          },
          "401": { "$ref": "#/components/responses/AuthenticationRequiredResponse" },
          "404": { "$ref": "#/components/responses/NotFoundResponse" }
        }
      }
    },
    "/api/members": {
      "post": {
        "operationId": "createMember",
        "tags": ["Catalog"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/MemberCreateRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Membership registration accepted in a mutable environment",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MemberCreateResponse" } }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "405": { "$ref": "#/components/responses/ReadOnlyResponse" }
        }
      }
    },
    "/api/ingredientes": {
      "get": {
        "operationId": "getIngredients",
        "tags": ["Laboratory"],
        "parameters": [
          { "$ref": "#/components/parameters/IngredientId" },
          { "$ref": "#/components/parameters/IngredientFamily" },
          { "$ref": "#/components/parameters/IngredientStatus" },
          { "$ref": "#/components/parameters/SearchQuery" }
        ],
        "responses": {
          "200": {
            "description": "Ingredient detail when id is present, otherwise a filtered collection",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/IngredientDetail" },
                    { "$ref": "#/components/schemas/IngredientCollection" }
                  ]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundResponse" }
        }
      }
    },
    "/api/hipotesis": {
      "get": {
        "operationId": "listHypotheses",
        "tags": ["Laboratory"],
        "responses": {
          "200": {
            "description": "Technical-sheet registry",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/HypothesisCollection" } }
            }
          },
          "500": { "$ref": "#/components/responses/ServerErrorResponse" }
        }
      },
      "post": {
        "operationId": "createHypothesis",
        "tags": ["Laboratory"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/HypothesisCreateRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/HypothesisMutationResponse" },
          "201": { "$ref": "#/components/responses/HypothesisMutationResponse" },
          "400": { "$ref": "#/components/responses/ValidationErrorResponse" },
          "422": { "$ref": "#/components/responses/HypothesisValidationResponse" },
          "405": { "$ref": "#/components/responses/ReadOnlyResponse" },
          "500": { "$ref": "#/components/responses/ServerErrorResponse" }
        }
      }
    },
    "/api/hipotesis/{id}": {
      "get": {
        "operationId": "getHypothesis",
        "tags": ["Laboratory"],
        "parameters": [{ "$ref": "#/components/parameters/HypothesisId" }],
        "responses": {
          "200": {
            "description": "One technical sheet",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/HypothesisDetail" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundResponse" },
          "500": { "$ref": "#/components/responses/ServerErrorResponse" }
        }
      }
    },
    "/api/test/reset": {
      "post": {
        "operationId": "resetTestData",
        "tags": ["Test support"],
        "description": "Test-only hook. Development and production conceal it with 404.",
        "responses": {
          "200": {
            "description": "Test fixture reset acknowledgement",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ResetResponse" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundResponse" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lumbre_auth.session_token"
      }
    },
...
    "responses": {
...
      "ValidationErrorResponse": {
        "description": "Request validation failed",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
...
      "AdministrativeProductCollectionResponse": {
        "description": "Complete product catalog visible to administrators",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/AdministrativeProductCollection" } }
        }
      },
...
      "AuthenticationRequiredResponse": {
        "description": "An authenticated account is required",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "AuthorizationRequiredResponse": {
        "description": "The authenticated account lacks the required role",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "AuthenticationUnavailableResponse": {
        "description": "Authentication is not configured for the active environment",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "MembershipPreferencesResponse": {
        "description": "The authenticated account's cooking preferences and consent history",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPreferencesResponse" } }
        }
      },
      "MembershipPreferenceMutationResponse": {
        "description": "Created or updated account preferences with consent audit metadata",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/MembershipPreferenceMutation" } }
        }
      },
...
    },
    "schemas": {
...
      "AccountResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "oneOf": [
              { "$ref": "#/components/schemas/Account" },
              { "type": "null" }
            ]
          }
        },
        "additionalProperties": false
      },
      "AccountCollection": {
        "type": "object",
        "required": ["data", "count"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Account" } },
          "count": { "type": "integer", "minimum": 0 }
        },
        "additionalProperties": false
      },
      "MagicLinkRequest": {
        "type": "object",
        "required": ["name", "email"],
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 100 },
          "email": { "type": "string", "format": "email", "maxLength": 254 }
        },
        "additionalProperties": false
      },
      "MagicLinkResponse": {
        "type": "object",
        "required": ["status"],
        "properties": { "status": { "const": true } },
        "additionalProperties": false
      },
      "LogoutResponse": {
        "type": "object",
        "required": ["success"],
        "properties": { "success": { "const": true } },
        "additionalProperties": false
      },
...
      "MembershipPreferenceUpdateRequest": {
        "$ref": "#/components/schemas/MembershipPreferenceInput"
      },
...
    }
  }
}